+++ lib/resources/customer.ts
@@ -1,4 +1,4 @@
-const {ApiResource} = require("./api");
+import ApiResource from "./api";
 
 class Customer extends ApiResource {
   constructor(appId?: string, secretKey?: string) {
@@ -23,29 +23,42 @@ class Customer extends ApiResource {
 
   async resetApiKey(customer_uid: string, version?: number) {
     const path = `/api/v${version}/${this._objectName}/${customer_uid}/reset-api-key`;
-    return this._requestor.request('POST', path);
+    return this.postRequest(path);
   }
 
-  async createCheckoutSession(customer_uid: string, product_uid: string, version?: number) {
+  async createCheckoutSession(
+    customer_uid: string,
+    product_uid: string,
+    version?: number
+  ) {
     const path = `/api/v${version}/${this._objectName}/${customer_uid}/checkout-session/${product_uid}`;
-    return this._requestor.request('POST', path);
+    return this.postRequest(path);
   }
 
-  async createSandboxSubscription(customer_uid: string, product_uid: string, data: any, version?: number) {
+  async createSandboxSubscription(
+    customer_uid: string,
+    product_uid: string,
+    data: any,
+    version?: number
+  ) {
     const path = `/api/v${version}/${this._objectName}/${customer_uid}/create-promo/${product_uid}`;
-    return this._requestor.request('POST', path, data);
+    return this.postRequest(path);
   }
 
   async cancelSubscription(customer_uid: string, version?: number) {
     const path = `/api/v${version}/${this._objectName}/${customer_uid}/cancel-subscription`;
-    return this._requestor.request('POST', path);
+    return this.postRequest(path);
   }
 
-  async createPaymentLink(customer_uid: string, product_id: string, data: any, version?: number) {
+  async createPaymentLink(
+    customer_uid: string,
+    product_id: string,
+    data: any,
+    version?: number
+  ) {
     const path = `/api/v${version}/${this._objectName}/${customer_uid}/payment-link/${product_id}`;
-    return this._requestor.request('POST', path, data);
+    return this.postRequest(path);
   }
-
 }
 
-module.exports = Customer;
+export default Customer;